# Refactor: Remove background prefetch of schedules
This commit removes the feature that automatically downloaded all schedules in the background upon app launch. This was causing unnecessary
 network usage and potential performance issues.

The setting to enable this feature has been removed from the settings screen and the corresponding logic has been deleted from the Repository class.

Additionally, error handling has been improved in the App class to catch IllegalStateException when navigating and report it to Firebase Crashlytics.

Finally, the Gradle wrapper has been updated to version 8.7.